\name{bioCarta.plot}
\alias{bioCarta.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plots a graph of the selected bioCarta pathway
}
\description{
This function plots a graph of the selected bioCarta pathway.
}
\usage{
bioCarta.plot(pid, ratios = NULL, species = "Hs")
}
\arguments{
  \item{pid}{a numeric specifying the pathway id of which bioCarta pathway to generate.}
  \item{ratios}{a named numeric vector specifying the gene expression values to overlay on gene nodes.
 Names must correspond to probes from the the \code{chipAnnotation} specified chip.}
  \item{species}{default = "Hs", human data. Mouse ratios will also work using "Mm"}
}
\details{
This function generates a graph of the specified bioCarta pathway. \cr
Nodes represent either molecular interactions that indicated as white boxes. These connect
 between various molecule nodes (light blue, ellipse). \cr
Edges represent either positive (green) or negative (red) interactions.

Each molecular nodes may represent an individual or groups of genes.
 If \code{ratios} is specified then the gene nodes are colored according to expresion
 value. Multiple genes are represented by segments in a pie chart.

Ratios should be named numeric vector. The names correspond to probe or \code{featureNames} 
for the given microarray.\cr
The nodes are colored as follows:\cr
\tabular{rrrrrrrr}{
 \bold{Color =} \tab darkgreen, \tab green, \tab palegreen,
 \tab lightyellow, \tab orange, \tab red, \tab darkred\cr
 \bold{Ratio =} \tab >-2, \tab -2 to -1, \tab -1 to -0.5, \tab -0.5 to 0.5, 
 \tab 0.5 to 1, \tab 1 to 2, \tab >2
}

The numeric ratio is the ratio to be ploted in nodes of the bioCarta graph generate. Typically, these 
are taken from contrast coefficients generated by limma, or manually generated by comparing two groups.

Currently ratio plotting only works with mouse and human data. 

The output of these graphs can be quite large and is often best exported as a PDF file so that one can
zoom in to read small labels on complex pathways.
}
\value{
This function returns nothing. It is called only for the side effect of
 producing the bioCarta graph of the specified pathway.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Fraser Sim <fjsim@buffalo.edu>
}
%\note{
%%  ~~further notes~~
%}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
 \code{\link{bioCarta.heatmap}}, \code{\link{KEGG.plot}}
}
\examples{
 bioCarta.plot(10) 

 data(bioCarta)
 print(PathwayNames) # lists all available bioCarta pathways

 # load data
 data(eset)
 # set chipAnnotation
 chipAnnotation = annotation(eset)

 # use limma to define differentially expressed transcipts
 # refer to limma documentation for details on linear model design.
 library(limma)
 design = model.matrix(~-1 + eset$phenotype)
 colnames(design) <- levels(eset$phenotype)
 fit <- lmFit(eset, design)

 # contrast.matrix may be set up using multiple contrasts
 contrast.matrix <- makeContrasts(A2B5 - uns, levels = design)
 
 fit2 <- contrasts.fit(fit, contrast.matrix)
 fit3 <- eBayes(fit2)

 # pdgf signal pathway with ratios from sample data
 bioCarta.plot(141, ratios = fit3$coefficients[,1])
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
