\name{dataSets}
\alias{eset}
\alias{contrast.details}
\docType{data}
\title{
Example data taken from Sim et al., 2006
}
\description{
The example data used in this package is publically available at NCBI GEO - GSE26535.
}
\usage{data(eset)}
%\format{
%}
\details{
eset = \code{ExpressionSet} object generated by running \code{\link[affy]{justRMA}} \cr
contrast.details = a list object generated by performing limma analysis and \link{parseContrasts}
}
\source{
NCBI GEO - GSE26535 - \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE26535}
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(eset)
data(fit3)
data(contrast.details)
}
\keyword{datasets}
