\name{KEGG.plot}
\alias{KEGG.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plots a graph of the selected KEGG pathway
}
\description{
This function plots a graph of the selected KEGG pathway.
}
\usage{
KEGG.plot(KeggID, ratios = NULL, chipAnnotation = "hgu133plus2", expandGenes = FALSE, species = "Hs")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{KeggID}{a character specifying the pathway id of which bioCarta pathway to generate.}
  \item{ratios}{a named numeric vector specifying the gene expression values to overlay on gene nodes.}
  \item{chipAnnotation}{see other}
  \item{expandGenes}{a \code{boolean} specifying whether or not to expand nodes of related genes.}
  \item{species}{a \code{character} variable specifying the species of the expression set data.\cr
   Only 'Hs' and 'Mm' are currently supported.}
}
\details{
This function generates a graph representation of a selected KEGG pathway. If \code{ratios} is specified,
 individual or grouped gene nodes are colored according to an expression ratio.

Using \code{\link[KEGGgraph]{parseKGML}} & \code{\link[KEGGgraph]{KEGGpathway2Graph}} - an individual
 KEGG pathway is downloaded and parsed into a graph object. The graph object is then plotted using
 \code{\link[Rgraphviz]{renderGraph}}.

The nodes represent individual genes (light blue ellipse) or compounds (purple square).
The edges represent interactions between nodes and are colored and labelled as follows:
\tabular{lll}{
 \bold{edgeType} \tab \bold{color} \tab \bold{arrowhead} \cr
 activation \tab green \tab normal \cr
 activation_indirect \tab palegreen \tab normal \cr
 activation_phosphorylation \tab green \tab normal \cr
 inhibition \tab red \tab tee \cr
 inhibition_phosphorylation \tab red \tab tee \cr
 inhibition_dephosphorylation \tab red \tab tee \cr
 compound \tab purple \tab diamond \cr
 expression \tab palegreen \tab normal \cr
 indirect \tab grey \tab normal \cr
 phosphorylation \tab grey \tab normal \cr
 dephosphorylation \tab grey \tab normal \cr
 ubiquitination \tab orange \tab diamond \cr
 activation_ubiquination \tab orange \tab diamond \cr
 ubiquination_inhibition \tab orange \tab diamond \cr
 binding/association \tab yellow \tab dot \cr
 phosphorylation_binding/association \tab yellow \tab dot \cr
 dissociation \tab cyan \tab odot \cr
 missing \tab grey \tab none 
}

Ratios should be named numeric vector. The names correspond to probe or \code{featureNames} 
for the given microarray.\cr
The nodes are colored as follows:\cr
\tabular{rrrrrrrr}{
 \bold{Color =} \tab darkgreen, \tab green, \tab palegreen,
 \tab lightyellow, \tab orange, \tab red, \tab darkred\cr
 \bold{Ratio =} \tab >-2, \tab -2 to -1, \tab -1 to -0.5, \tab -0.5 to 0.5, 
 \tab 0.5 to 1, \tab 1 to 2, \tab >2
}

Typically, ratio values are taken from contrast coefficients generated by limma, or
 manually generated by comparing two groups.

Currently ratio plotting only works with mouse and human data. 

The output of these graphs can be quite large and is often best exported as a PDF file so that one can
zoom in to read small labels on complex pathways.
}
\value{
If \code{ratios} specified, then the function returns a \code{character} vector of probe sets in pathway.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Fraser Sim <fjsim@buffalo.edu>
}
%\note{
%%  ~~further notes~~
%}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
 \code{\link{bioCarta.plot}}
}
\examples{
# To do
\donttest{KEGG.plot("04340") # plot KEGG pathway "Hedgehog signaling pathway"}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
