\name{mouseOPC}
\alias{mouseOPC}
\docType{data}
\title{
Example data taken from Cahoy et al., 2008
}
\description{
The example data is taken from Supplemental Table 17 of Cahoy et al., 2008.
}
\usage{data(eset)}
%\format{
%}
\details{
A character vector of significantly enriched probe sets in mouse oligodendrocyte progenitor cells.
}
\source{
Cahoy JD, Emery B, Kaushal A, Foo LC, Zamanian JL, Christopherson KS, Xing Y,
 Lubischer JL, Krieg PA, Krupenko SA, Thompson WJ, Barres BA (2008)
 A transcriptome database for astrocytes, neurons, and oligodendrocytes:
 a new resource for understanding brain development and function. J Neurosci 28:264-278.
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(eset)
data(fit3)
data(contrast.details)
}
\keyword{datasets}
