\name{results}
\alias{results.topGO}
\alias{results.KEGG}
\docType{data}
\title{
Example results of topGO and KEGG pathway analysis
}
\description{
This example data was generated using the example ExpressionSet and runAutomatedHyperG
}
\usage{data(eset)}
%\format{
%}
\details{
In order to improve speed of vignette and package build, the results of runAutomatedHyperG were exported
as a Rda file.
}
%\source{
%}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(results)
}
\keyword{datasets}
