\name{runAutomatedGOSummary}
\alias{runAutomatedGOSummary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
run automated GO summary on all contrasts 
}
\description{
This function categorizes differentially expressed genes in common subtypes according to GO annotations.
}
\usage{
runAutomatedGOSummary(contrast.details)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{contrast.details}{a \code{list} object generated by \code{\link{parseContrasts}}.}
}
\details{
This function generates csv files (comma delimited data tables) for each contrast.\cr
Differentially expressed genes are first categorized into ligands, receptors, transcription factors, extra-
cellular matrix (ECM), and enzyme genes. A table is then constructed and exported for each of the up- and 
down-regulated gene lists defined in \code{contrast.details}.
}
\value{
This function does not return any value. A series of csv files are generated in the working directory.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Fraser Sim <fjsim@buffalo.edu>
}
%\note{
%%  ~~further notes~~
%}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
 \code{\link{parseContrasts}}, \code{\link{GOSummary}}
}
\examples{
# load data
data(eset)
# set chipAnnotation
chipAnnotation = annotation(eset)

# load contrast.details
data(contrast.details)

runAutomatedGOSummary(contrast.details)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
